core/requirements.txt
asgiref==3.7.2
Django==4.2
django-environ==0.10.0
django-filter==23.2
djangorestframework==3.14.0
drf-writable-nested==0.7.0
Pillow==9.5.0
psycopg2-binary==2.9.6
Pygments==2.15.1
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
typing_extensions==4.6.3
uritemplate==4.1.1
